[MISRA-C] apply MISRA-C:2012 Rule 20.7 - parenthesize macro parameters
Enclose macro parameters module_id and is_enabled in parentheses in
the MODULE_CONFIG() macro in device_tree.h (aiacc, mdp_x5h, x5h) to
prevent unintended operator precedence issues.

+++ FreeRTOS/Demo/R-Car_Gen5_CR52/common/include/r_pfc_common_private.h
@@ -80,7 +80,7 @@ typedef enum e_modsel_func
 } modsel_func_t;
 
 #define GEN_ID(reg_pfc_t, grp, pin, fid)   \
-((reg_pfc_t<<REG_TYPE_POS) + (grp<<GROUP_START_POS) + (pin<<PIN_START_POS) + (fid<<FUNC_START_POS))
+(((uint32_t)(reg_pfc_t)<<REG_TYPE_POS) + ((uint32_t)(grp)<<GROUP_START_POS) + ((uint32_t)(pin)<<PIN_START_POS) + ((uint32_t)(fid)<<FUNC_START_POS))
 
 /******************* Define pin functions *******************/
 #define MAX_ITEM_IN_GROUP   16
